update for BCOrgSalaryStatisticsReport

+++ work/src/org/kuali/kfs/module/bc/businessobject/BudgetConstructionOrgSalaryStatisticsReport.java
@@ -42,9 +42,9 @@ public class BudgetConstructionOrgSalaryStatisticsReport {
     private BigDecimal appointmentRequestedFteQuantity;
     private Integer totalCsfAmount;
     private Integer totalAppointmentRequestedAmount;
-    private Integer averageCsfAmount;
-    private Integer averageAppointmentRequestedAmount;
-    private Integer averageChange;
+    private BigDecimal averageCsfAmount;
+    private BigDecimal averageAppointmentRequestedAmount;
+    private BigDecimal averageChange;
     private BigDecimal percentChange;
     
     
@@ -99,22 +99,22 @@ public class BudgetConstructionOrgSalaryStatisticsReport {
     public void setOrganizationName(String organizationName) {
         this.organizationName = organizationName;
     }
-    public Integer getAverageAppointmentRequestedAmount() {
+    public BigDecimal getAverageAppointmentRequestedAmount() {
         return averageAppointmentRequestedAmount;
     }
-    public void setAverageAppointmentRequestedAmount(Integer averageAppointmentRequestedAmount) {
+    public void setAverageAppointmentRequestedAmount(BigDecimal averageAppointmentRequestedAmount) {
         this.averageAppointmentRequestedAmount = averageAppointmentRequestedAmount;
     }
-    public Integer getAverageChange() {
+    public BigDecimal getAverageChange() {
         return averageChange;
     }
-    public void setAverageChange(Integer averageChange) {
+    public void setAverageChange(BigDecimal averageChange) {
         this.averageChange = averageChange;
     }
-    public Integer getAverageCsfAmount() {
+    public BigDecimal getAverageCsfAmount() {
         return averageCsfAmount;
     }
-    public void setAverageCsfAmount(Integer averageCsfAmount) {
+    public void setAverageCsfAmount(BigDecimal averageCsfAmount) {
         this.averageCsfAmount = averageCsfAmount;
     }
     public BigDecimal getInitialRequestedFteQuantity() {
